+++ Burger-Queen-Api-Client/src/app/services/AdminService/admin.service.ts
@@ -29,16 +29,7 @@ export class AdminService {
   // HACER PETICIÓN PETICIÓN POST..
   // MANDO BODY MI OBJETO CON TODA LA DATA PARA CREAR UNA ORDEN
   // ES NECESARIO.. CONECTAR A LA BASE DE DATOS!!!
-  // getListProducts(id: number, name: string, price:number, image: string, type:string, dateEntry:string): Observable<any> {
-  //   return this.http.post('http://localhost:8080/users/products', {
-  //     id:id
-      // name: name,
-  //     price: price,
-  //     image: image,
-  //     type: type,
-      // dateEntry: dateEntry,
-  //   })
-  // })
-
-
+  getListProducts(): Observable<any> {
+    return this.http.get('http://localhost:8080/products',)
+  }
   }
